chore: move useSyncFormParameters to shared hook
